+++ nightingale_dispatcher/src/nightingale_dispatcher/subtask.py
@@ -1,0 +1,13 @@
+class Subtask:
+    SUCCESS = 0
+    ARGUMENT_ERROR = 1
+    SERVICE_ERROR = 2
+    TOPIC_ERROR = 3
+    ACTION_ERROR = 4
+
+    def __init__(self, name, priority=0):
+        self.name = name
+        self.priority = priority
+
+    def execute(self):
+        pass
