+++ nightingale_msgs/srv/AddTaskDispatcher.srv
@@ -1,0 +1,4 @@
+string name
+string items_json
+---
+int8 status
