+++ nightingale_msgs/srv/CancelTaskDispatcher.srv
@@ -1,0 +1,3 @@
+string name
+---
+int8 status
